Engine folder for audio decoding and playback.
