+++ src/app/pages/apps/pages/audio/components/engine/interfaces.ts
@@ -1,0 +1,10 @@
+export interface PlaybackState {
+    state: string;
+}
+
+export interface DecodedData {
+    channelData: Array<Float32Array>;
+    length: number;
+    numChannels: number;
+    sampleRate: number;
+}
